+++ client-ts/package.json
@@ -10,14 +10,17 @@
     "eject": "react-scripts eject"
   },
   "dependencies": {
+    "@date-io/date-fns": "^2.5.0",
     "@google/markerclusterer": "^2.0.6",
     "@material-ui/core": "^4.9.4",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "^4.0.0-alpha.44",
+    "@material-ui/pickers": "^3.2.10",
     "axios": "^0.19.2",
     "chart.js": "^2.9.3",
     "chartjs-plugin-colorschemes": "^0.4.0",
     "classnames": "^2.2.6",
+    "date-fns": "^2.12.0",
     "dotenv": "^8.2.0",
     "history": "^4.10.1",
     "html2canvas": "^1.0.0-rc.5",
