+++ client-ts/src/organisms/mypage/marketer/dashboard/CampaignAnalysisDialogV2.tsx
@@ -39,7 +39,7 @@ export default function CampaignAnalysisDialogV2({
   );
 
   const creatorsData = useGetRequest<{ campaignId: string }, CreatorDataInterface[] | null>(
-    '/marketer/campaign/analysis/v1/creator-data',
+    '/marketer/campaign/analysis/creator-data',
     { campaignId: selectedCampaign.campaignId }
   );
