+++ server/routes/marketer/campaign/analysis_v2.ts
@@ -84,4 +84,75 @@ router.route('/heatmap')
   .all(responseHelper.middleware.unusedMethod);
 
 
+// 마케터 캠페인 분석에 사용되는 크리에이터 데이터
+// marketer/sub/report =>/creators
+router.route('/creator-data')
+  .get(
+    responseHelper.middleware.checkSessionExists, // session 확인이 필요한 경우.
+    responseHelper.middleware.withErrorCatch(async (req, res, next) => {
+      const { marketerId } = responseHelper.getSessionData(req);
+      const campaignId = responseHelper.getParam('campaignId', 'GET', req);
+      let query = '';
+      let queryArray = [];
+      if (campaignId !== '') {
+        query = `
+        SELECT
+        ci.creatorId AS creatorId, 
+        ci.creatorName AS creatorName, 
+        ci.creatorTwitchId AS creatorTwitchId,
+        ci.creatorLogo AS creatorLogo, 
+        sum(cashFromMarketer) AS total_ad_exposure_amount,
+        cd.viewer AS viewer,
+        cd.followers AS followers, 
+        cd.airtime AS airtime, 
+        cd.impression AS impression, 
+        cd.openHour AS openHour, 
+        cd.content AS content
+        FROM campaignLog as cl
+        JOIN creatorInfo as ci
+        ON cl.creatorId = ci.creatorId
+        LEFT JOIN creatorDetail AS cd
+        ON cl.creatorId = cd.creatorId
+        WHERE campaignId = ?
+        GROUP BY cl.creatorId
+        ORDER BY total_ad_exposure_amount DESC`;
+
+        queryArray = [campaignId];
+      } else {
+        query = `
+        SELECT
+        ci.creatorId AS creatorId, 
+        ci.creatorName AS creatorName, 
+        ci.creatorTwitchId AS creatorTwitchId,
+        ci.creatorLogo AS creatorLogo, 
+        sum(cashFromMarketer) AS total_ad_exposure_amount,
+        cd.viewer AS viewer,
+        cd.followers AS followers, 
+        cd.airtime AS airtime, 
+        cd.impression AS impression, 
+        cd.openHour AS openHour, 
+        cd.content AS content
+        FROM campaignLog as cl
+        JOIN creatorInfo as ci
+        ON cl.creatorId = ci.creatorId
+        LEFT JOIN creatorDetail AS cd
+        ON cl.creatorId = cd.creatorId
+        WHERE SUBSTRING_INDEX(campaignId, '_', 1) = ?
+        GROUP BY cl.creatorId
+        ORDER BY total_ad_exposure_amount DESC`;
+
+        queryArray = [marketerId];
+      }
+      doQuery(query, queryArray)
+        .then((row) => {
+          responseHelper.send(row.result, 'get', res);
+        })
+        .catch((error) => {
+          responseHelper.promiseError(error, next);
+        });
+    }),
+  )
+  .all(responseHelper.middleware.unusedMethod);
+
+
 export default router;
